+++ R/utils-plumber.R
@@ -431,3 +431,23 @@ citation_from_version <- function(version) {
   )
   )
 }
+
+#' create_etag_header
+#'
+#' helper function that creates a unique hash of code + data
+#' this hash value will be used as the value of the etag header
+#' to facilitate caching of PIP API responses
+#'
+#' @param req R6 object: Plumber API request
+#'
+#' @return character
+
+create_etag_header <- function(req){
+  lkup_hash   <- lkups$versions_paths[[req$argsQuery$version]]
+  pipapi_hash <- packageDescription("pipapi")$GithubSHA1
+  wbpip_hash  <- packageDescription("wbpip")$GithubSHA1
+
+  etag_hash <- rlang::hash(c(lkup_hash, pipapi_hash, wbpip_hash))
+
+  return(etag_hash)
+}