+++ NEWS.md
@@ -1,4 +1,4 @@
-# pipapi (development version)
+# pipapi 1.2.1
 
 ## New features
 - New indicators available. Optional with `additional_ind = TRUE`
@@ -10,6 +10,10 @@
 - A new `/ui_aux` endpoint has been created with the opposite behavior: tables are
 returned in wide format by default.
 - `/pip-info` returns additional information
+- New "etag" header returned by the API to facilitate caching of API responses
+
+## Enhancements
+- pip-grp and cp-chart endpoint small optimization
 
 ## Enhancements